requirements.txt
boto3==1.28.65
botocore==1.31.65
click==8.1.7
httpx==0.27.2
prometheus-client==0.21.0
pbr==6.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.9
flake8==6.1.0
bandit==1.7.5
